reuse parsing code for version numbers, remove parseopt.h/cpp
