--- src/commandline/parseopt.h
@@ -1,7 +1,0 @@
-#if !defined(_RSYS_PARSEOPT_H_)
-#define _RSYS_PARSEOPT_H_
-namespace Executor
-{
-extern bool parse_version(std::string vers, uint32_t *version_out);
-}
-#endif /* !_RSYS_PARSEOPT_H_ */
